Refactor the form from "Edit recipe" page into EditRecipeForm component.
